Compatibility with Polylang 3.4.3 (#2)
This PR fixes the issue where, since PLL 3.4.3, the info related to the
DB are returned by `PLL_Translated_Object::get_db_infos()` instead of
being stored in `PLL_Translated_Object::$db`. This was causing a fatal
error because of the abstract method not being implemented.

Backward compatibility with the `$db` property is assured.

This PR also bumps the plugin's version to 1.1.